refactor: subscribeQuestion 관련 클래스 subscribe 내부로 이동
