+++ src/main/java/maeilmail/statistics/EventAggregator.java
@@ -3,7 +3,7 @@ package maeilmail.statistics;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
-import maeilmail.subscribequestion.SubscribeQuestion;
+import maeilmail.subscribe.command.domain.SubscribeQuestion;
 import org.springframework.stereotype.Component;
 
 @Component
