+++ src/main/java/maeilmail/subscribe/query/SubscribeQuestionQueryService.java
@@ -1,4 +1,4 @@
-package maeilmail.subscribequestion;
+package maeilmail.subscribe.query;
 
 import static maeilmail.question.QQuestion.question;
 import static maeilmail.subscribe.command.domain.QSubscribe.subscribe;
@@ -11,6 +11,8 @@ import com.querydsl.jpa.impl.JPAQuery;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
 import maeilmail.question.QuestionCategory;
+import maeilmail.subscribequestion.QSubscribeQuestionSummary;
+import maeilmail.subscribequestion.QWeeklySubscribeQuestionSummary;
 import maeilmail.support.DateUtils;
 import maeilmail.support.PaginationResponse;
 import org.springframework.data.domain.Page;
