+++ src/main/java/maeilmail/subscribe/query/SubscribeQuestionSummary.java
@@ -1,4 +1,4 @@
-package maeilmail.subscribequestion;
+package maeilmail.subscribe.query;
 
 import com.querydsl.core.annotations.QueryProjection;
